We now really do have all the needed framework for the semisimplicial type presented by a relational semisimplicial complex.
We have tried to avoid even mentioning equality, for the sake of having
proof terms with good behaviour under partial evaluation; part of this is
explicitly constructing the coinhabitation arguments that would be provided
by transport [ and which, with Funext and Univalence, conversely provide
the identification to transport over ]

There are, however, one or two spots where transport is used, though I'm
sure these could be eliminated as well, and I don't know what effect
keeping them will have on the next layers of development.

lList has been adjusted so as to more-closely resemble nCkList; I don't
know if this is important.

nCkEqns does have equations in it, but it is mostly about most of the
coinhabitations mentioned above.